# Optimize toByteString and toASCIIBytes