--- a/CONTRIBUTORS
+++ b/CONTRIBUTORS
@@ -13,3 +13,4 @@ Michael Snoyman
 Bardur Arantsson
 Timo von Holtz
 Oleg Grenrus
+Borys Lykah