--- a/uuid-bench/uuid-bench.cabal
+++ b/uuid-bench/uuid-bench.cabal
@@ -40,19 +40,19 @@ source-repository head
 
 library
   build-depends:
-      base              >=4.9      && <5
-    , binary            >=0.5.1.0  && <0.9
-    , bytestring        >=0.9.2.1  && <0.12
-    , cryptohash-md5    >=0.11.100 && <0.12
-    , cryptohash-sha1   >=0.11.100 && <0.12
-    , deepseq           >=1.3.0.0  && <1.5
-    , entropy           >=0.3.7    && <0.5
-    , hashable          >=1.2.7.0  && <1.5
-    , network-info      >=0.2      && <0.3
-    , random            >=1.1      && <1.3
-    , template-haskell  >=2.7.0.0  && <2.20
-    , text              >=1.2.3.0  && <1.3 || >=2.0 && <2.1
-    , time              >=1.4      && <1.13
+      base              >=4.9       && <5
+    , binary            >=0.5.1.0   && <0.9
+    , bytestring        >=0.10.2.0  && <0.12
+    , cryptohash-md5    >=0.11.100  && <0.12
+    , cryptohash-sha1   >=0.11.100  && <0.12
+    , deepseq           >=1.3.0.0   && <1.5
+    , entropy           >=0.3.7     && <0.5
+    , hashable          >=1.2.7.0   && <1.5
+    , network-info      >=0.2       && <0.3
+    , random            >=1.1       && <1.3
+    , template-haskell  >=2.7.0.0   && <2.21
+    , text              >=1.2.3.0   && <1.3 || >=2.0 && <2.1
+    , time              >=1.4       && <1.13
 
   -- uuid modules
   exposed-modules: