--- a/uuid-types/uuid-types.cabal
+++ b/uuid-types/uuid-types.cabal
@@ -50,14 +50,14 @@ source-repository head
 
 library
   build-depends:
-      base              >=4.5     && <5
-    , binary            >=0.5.1.0 && <0.9
-    , bytestring        >=0.9.2.1 && <0.13
-    , deepseq           >=1.3.0.0 && <1.6
-    , hashable          >=1.2.7.0 && <1.5
-    , random            >=1.1     && <1.3
-    , template-haskell  >=2.7.0.0 && <2.22
-    , text              >=1.2.3.0 && <1.3  || >=2.0 && <2.2
+      base              >=4.5      && <5
+    , binary            >=0.5.1.0  && <0.9
+    , bytestring        >=0.10.2.0 && <0.13
+    , deepseq           >=1.3.0.0  && <1.6
+    , hashable          >=1.2.7.0  && <1.5
+    , random            >=1.1      && <1.3
+    , template-haskell  >=2.7.0.0  && <2.22
+    , text              >=1.2.3.0  && <1.3  || >=2.0 && <2.2
 
   exposed-modules:  Data.UUID.Types