--- a/uuid/uuid.cabal
+++ b/uuid/uuid.cabal
@@ -1,6 +1,6 @@
 cabal-version:      1.12
 name:               uuid
-version:            1.3.15
+version:            1.3.16
 x-revision:         3
 copyright:          (c) 2008-2014 Antoine Latter
 author:             Antoine Latter
@@ -46,16 +46,16 @@ source-repository head
 
 library
   build-depends:
-      base             >=4.5      && <5
-    , binary           >=0.5.1.0  && <0.9
-    , bytestring       >=0.9.2.1  && <0.13
-    , cryptohash-md5   >=0.11.100 && <0.12
-    , cryptohash-sha1  >=0.11.100 && <0.12
-    , entropy          >=0.3.7    && <0.5
-    , network-info     >=0.2      && <0.3
-    , random           >=1.1      && <1.3
-    , text             >=1.2.3.0  && <1.3 || >=2.0 && <2.2
-    , time             >=1.4      && <1.13
+      base             >=4.5       && <5
+    , binary           >=0.5.1.0   && <0.9
+    , bytestring       >=0.10.2.0  && <0.13
+    , cryptohash-md5   >=0.11.100  && <0.12
+    , cryptohash-sha1  >=0.11.100  && <0.12
+    , entropy          >=0.3.7     && <0.5
+    , network-info     >=0.2       && <0.3
+    , random           >=1.1       && <1.3
+    , text             >=1.2.3.0   && <1.3 || >=2.0 && <2.2
+    , time             >=1.4       && <1.13
 
   -- strict dependency on uuid-types,
   -- as we re-rexport datatype, thus leak instances etc.