@lykahb 2024-03-19
It is hard to tell what makes the high-level implementation faster. The `toByteString` packed the characters from a list, which looks suboptimal. But the `toASCIIBytes` only does the low-level operations and looks quite simple, with, as you said, hard to imagine anything being faster.

Yet, the benchmarks show that a high-level implementation is faster. It might be that it unlocks some GHC optimizations or that bytestring has some other non-trivial optimizations. If we compare only the functions that do not have any extra overhead (`toASCIIBytes` from before with `toByteString` in this PR), it'd be a third faster.

If `uuid` delegates the low-level work to `bytestring`, it would benefit from any future performance improvements in `bytestring`. Also, this might be easier to maintain: this diff mostly removes code and lets this library focus on the format, rather than the low-level optimization.

There is also more room for speedup: `toASCIIBytes` can be as just fast as `toByteString` if we remove the copy https://github.com/haskell/bytestring/issues/666. For better compatibility I avoided using an internal module in this PR `Data.ByteString.Builder.Prim.Internal.runF`.