@clyring 2024-03-19
If I had to guess, the reason the `FixedPrim` implementation in this patch is faster is probably to do with `BBP.word32HexFixed` being a branchless two-nibbles-at-a-time lookup instead of the branchy one-nibble-at-a-time calculation in the baseline.