@lykahb 2024-03-19
@clyring [clarified](https://github.com/haskell/bytestring/issues/666#issuecomment-2005704676) that the module `Data.ByteString.Builder.Prim.Internal` is stable. I rewrote the string conversion without the `Builder` and got further improvement:

```
benchmarking conversion/toASCIIBytes
time                 21.56 ns   (21.18 ns .. 22.30 ns)
                     0.995 R²   (0.988 R² .. 1.000 R²)
mean                 21.36 ns   (21.20 ns .. 21.95 ns)
std dev              968.8 ps   (101.2 ps .. 2.048 ns)


benchmarking conversion/toByteString
time                 17.85 ns   (17.82 ns .. 17.89 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 17.87 ns   (17.84 ns .. 17.95 ns)
std dev              166.1 ps   (89.32 ps .. 297.0 ps)
```