@lykahb 2024-03-20
With the recent update this branch has `toASCIIBytes` that is x1.6 faster and `toByteString` that is x4.15 faster.

@phadej Would you give this another look? This is a significant improvement and it reduces the amount of code to maintain in `uuid`. It seems that high-performance low-level code better fits the core libraries of the Haskell ecosystem like `bytestring`.