@phadej 2024-03-20
@iand675 the original patch was more complicated. And my review comment made lykahb improve on it.

Your comment is not fair.