@phadej 2024-03-20
@iand675 

> "complicate the code" 

If we don't understand why code works as it does, does it work?

In my opinion *it doesn't* in the long term. I'm the one who maintaining this, so I'm making the judgment calls.

This PR has a lot of good in it, but it's not perfect, I'll take if from here.

Note to self: benchmark on x86_64