@lykahb 2024-03-21
For the sake of clarity, is your objection about the use `Data.ByteString.Builder.Prim` in general or do you perceive anything else here as overly complicated? The only other thing I can think of was converting a `Builder` to a `ByteString` with `toLazyByteStringWith`. And that is gone now, inspired by Clyring's comment.